(#231)refactor: ContentType 공용 상수로 병합
