+++ src/main/java/com/adregamdi/like/dto/AllContentDTO.java
@@ -1,6 +1,6 @@
 package com.adregamdi.like.dto;
 
-import com.adregamdi.like.domain.enumtype.ContentType;
+import com.adregamdi.core.constant.ContentType;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class AllContentDTO{
+public class AllContentDTO {
 
     private String title;
     private ContentType contentType;
@@ -26,6 +26,7 @@ public class AllContentDTO{
 
     private String generateDetailUri(ContentType contentType, Long contentId) {
         return switch (contentType) {
+            case ALL -> null;
             case SHORTS -> "/api/shorts/stream/" + contentId;
             case PLACE -> "/api/places?place_id=" + contentId;
             case TRAVELOGUE -> "/api/travelogue?travelogue_id=" + contentId;
