--- src/main/java/com/adregamdi/notification/domain/ContentType.java
@@ -1,6 +1,0 @@
-package com.adregamdi.notification.domain;
-
-public enum ContentType {
-    SHORTS,
-    TRAVELOGUE
-}
