+++ src/main/java/com/adregamdi/place/application/PlaceServiceImpl.java
@@ -3,8 +3,8 @@ package com.adregamdi.place.application;
 import com.adregamdi.block.domain.Block;
 import com.adregamdi.block.exception.BlockException;
 import com.adregamdi.block.infrastructure.BlockRepository;
+import com.adregamdi.core.constant.ContentType;
 import com.adregamdi.like.application.LikesService;
-import com.adregamdi.like.domain.enumtype.ContentType;
 import com.adregamdi.media.application.ImageService;
 import com.adregamdi.member.domain.Member;
 import com.adregamdi.member.exception.MemberException;
